% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gelCoverage.r
\name{coverage_summary}
\alias{coverage_summary}
\title{A function to summarise coverage}
\usage{
coverage_summary(files, labels, covs, scope)
}
\arguments{
\item{files}{a single file or comma separated list of files}

\item{labels}{a single label or comma separated list of labels (same order as files)}

\item{covs}{ylim}

\item{scope}{wg or exome}
}
\description{
A function to summarise coverage
}
\examples{
coverage_summary()
}
\keyword{summary}

