% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gelCoverage.r
\name{cumulative_coverage}
\alias{cumulative_coverage}
\title{A function to plot cumulative coverage}
\usage{
cumulative_coverage(label, wgfile, exonfile, covs)
}
\arguments{
\item{label}{a single label i.e. well_id}

\item{wgfile}{the whole genome counts file}

\item{exonfile}{the exon counts file}

\item{covs}{ylim}
}
\description{
A function to plot cumulative coverage
}
\examples{
cumulative_coverage()
}
\keyword{cumulative}

