% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gelCoverage.r
\name{gc_exon_boxplots}
\alias{gc_exon_boxplots}
\title{A function to draw boxplots of coverage split by exon in transcript and gc content}
\usage{
gc_exon_boxplots(file)
}
\arguments{
\item{file}{a file of exon coverage and gc cotent}
}
\description{
A function to draw boxplots of coverage split by exon in transcript and gc content
}
\examples{
gc_exon_boxplots()
}
\keyword{boxplots}

