% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gelCoverage.r
\name{multiple_sample_plots}
\alias{multiple_sample_plots}
\title{Plots multiple samples on same graphs}
\usage{
multiple_sample_plots(files, labels, covs, scope)
}
\arguments{
\item{files}{a single files}

\item{labels}{a single label}

\item{covs}{ylim}

\item{scope}{wg or exome}
}
\description{
Plots multiple samples on same graphs
}
\examples{
multiple_sample_plots()
}
\keyword{summary}

