% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gelCoverage.r
\name{plot_gene_coverage}
\alias{plot_gene_coverage}
\title{A function to plot a gene with coverage and variants}
\usage{
plot_gene_coverage(gene, bw, vcf)
}
\arguments{
\item{gene}{the gene you want to print}

\item{bw}{the bw file}

\item{vcf}{the vcf file}
}
\description{
A function to plot a gene with coverage and variants
}
\examples{
plot_gene_coverage()
}
\keyword{plot}

