% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gelCoverage.r
\name{theme_gel_proper}
\alias{theme_gel_proper}
\title{A function to gel a plot}
\usage{
theme_gel_proper(base_size = 12, base_family = "Calibri")
}
\arguments{
\item{base_size}{base font size, default 12}

\item{base_family}{base font family, default Calibri}
}
\description{
A function to gel a plot
}
\examples{
theme_gel_proper()
}
\keyword{theme}

