% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gelCoverage.r
\name{coverage_summary}
\alias{coverage_summary}
\title{A function to summarise coverage}
\usage{
coverage_summary(file, label, covs, scope)
}
\arguments{
\item{covs}{ylim}

\item{scope}{wg or exome}

\item{files}{a single file}

\item{labels}{a single label}
}
\description{
A function to summarise coverage
}
\examples{
coverage_summary()
}
\keyword{summary}

